# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Toke Schools CBT Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/students:
    get:
      operationId: getAdminStudents
      tags: [admin]
      summary: List all students
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentRecord"
    post:
      operationId: createAdminStudent
      tags: [admin]
      summary: Create a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentRecord"
  /admin/students/{regNumber}:
    get:
      operationId: getAdminStudent
      tags: [admin]
      summary: Get a student by reg number
      parameters:
        - name: regNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentRecord"
    put:
      operationId: updateAdminStudent
      tags: [admin]
      summary: Update a student
      parameters:
        - name: regNumber
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: Student updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentRecord"
    delete:
      operationId: deleteAdminStudent
      tags: [admin]
      summary: Delete a student
      parameters:
        - name: regNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/staff:
    get:
      operationId: getAdminStaff
      tags: [admin]
      summary: List all staff
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffRecord"
    post:
      operationId: createAdminStaff
      tags: [admin]
      summary: Create a new staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffBody"
      responses:
        "201":
          description: Staff created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffRecord"
  /admin/staff/{staffId}:
    get:
      operationId: getAdminStaffMember
      tags: [admin]
      summary: Get a staff member by ID
      parameters:
        - name: staffId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Staff record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffRecord"
    put:
      operationId: updateAdminStaffMember
      tags: [admin]
      summary: Update a staff member
      parameters:
        - name: staffId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffBody"
      responses:
        "200":
          description: Staff updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffRecord"
    delete:
      operationId: deleteAdminStaffMember
      tags: [admin]
      summary: Delete a staff member
      parameters:
        - name: staffId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Staff deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/exams/{examId}:
    put:
      operationId: updateAdminExam
      tags: [admin]
      summary: Update any exam (admin only)
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExamBody"
      responses:
        "200":
          description: Exam updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
    delete:
      operationId: deleteAdminExam
      tags: [admin]
      summary: Delete any exam
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exam deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/exams/{examId}/results:
    get:
      operationId: getAdminExamResults
      tags: [admin]
      summary: Get all student results for any exam (admin only)
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResultWithStudent"
  /admin/exams/{examId}/results/{resultId}:
    delete:
      operationId: deleteAdminExamResult
      tags: [admin]
      summary: Remove a student's result (admin only), allows retake
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
        - name: resultId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Result removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
    StaffPermissions:
      type: object
      properties:
        manage_exams:
          type: boolean
        view_all_exams:
          type: boolean
        view_all_results:
          type: boolean
        manage_students:
          type: boolean
        reset_student_exam:
          type: boolean
      required: [manage_exams, view_all_exams, view_all_results, manage_students, reset_student_exam]
...
    StudentRecord:
      type: object
      properties:
        regNumber:
          type: string
        name:
          type: string
        class:
          type: string
        isDefaultPassword:
          type: boolean
        studentRole:
          type: string
          nullable: true
        profilePicture:
          type: string
          nullable: true
      required: [regNumber, name, class, isDefaultPassword]
    CreateStudentBody:
      type: object
      properties:
        regNumber:
          type: string
        name:
          type: string
        class:
          type: string
        password:
          type: string
        studentRole:
          type: string
          nullable: true
      required: [regNumber, name, class, password]
    UpdateStudentBody:
      type: object
      properties:
        name:
          type: string
        class:
          type: string
        password:
          type: string
          nullable: true
        resetPassword:
          type: boolean
          nullable: true
        studentRole:
          type: string
          nullable: true
      required: [name, class]
    StaffRecord:
      type: object
      properties:
        staffId:
          type: string
        name:
          type: string
        subject:
          type: string
        staffRole:
          type: string
        permissions:
          $ref: "#/components/schemas/StaffPermissions"
        profilePicture:
          type: string
          nullable: true
      required: [staffId, name, subject, staffRole, permissions]
    CreateStaffBody:
      type: object
      properties:
        staffId:
          type: string
        name:
          type: string
        subject:
          type: string
        staffRole:
          type: string
        permissions:
          $ref: "#/components/schemas/StaffPermissions"
        password:
          type: string
      required: [staffId, name, subject, staffRole, password]
    UpdateStaffBody:
      type: object
      properties:
        name:
          type: string
        subject:
          type: string
        staffRole:
          type: string
        permissions:
          $ref: "#/components/schemas/StaffPermissions"
        password:
          type: string
          nullable: true
      required: [name, subject, staffRole, permissions]
...
    Exam:
      type: object
      properties:
        id:
          type: integer
        subject:
          type: string
        class:
          type: string
        durationMinutes:
          type: integer
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        createdBy:
          type: string
        questionCount:
          type: integer
        alreadySubmitted:
          type: boolean
          nullable: true
        resultsEnabled:
          type: boolean
          nullable: true
      required: [id, subject, class, durationMinutes, createdBy, questionCount]
...
    CreateExamBody:
      type: object
      properties:
        subject:
          type: string
        class:
          type: string
        durationMinutes:
          type: integer
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
      required: [subject, class, durationMinutes]
...
    ResultWithStudent:
      type: object
      properties:
        id:
          type: integer
        studentReg:
          type: string
        studentName:
          type: string
        studentClass:
          type: string
        score:
          type: integer
        total:
          type: integer
        percentage:
          type: number
        grade:
          type: string
        submittedAt:
          type: string
      required: [id, studentReg, studentName, studentClass, score, total, percentage, grade, submittedAt]
...