# lib/api-zod/src/generated/types/createStaffBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Toke Schools CBT Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { StaffPermissions } from "./staffPermissions";

export interface CreateStaffBody {
  staffId: string;
  name: string;
  subject: string;
  staffRole: string;
  permissions?: StaffPermissions;
  password: string;
}