# lib/api-zod/src/generated/types/exam.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Toke Schools CBT Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Exam {
  id: number;
  subject: string;
  class: string;
  durationMinutes: number;
  startTime?: string | null;
  endTime?: string | null;
  createdBy: string;
  questionCount: number;
  alreadySubmitted?: boolean | null;
  resultsEnabled?: boolean | null;
}