# lib/api-zod/src/generated/types/studentRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Toke Schools CBT Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentRecord {
  regNumber: string;
  name: string;
  class: string;
  isDefaultPassword: boolean;
  studentRole?: string | null;
  profilePicture?: string | null;
}